Enable/Disable layers
Layers can be disabled. This stops its systems and entities from being updated. The current active entities are saved in the list activeEntities and are updated whenever a layer changes its status. The same applies for systems.
Moved addEntity, addSystem and deleteEntity from Layer to LayersEngine for better encapsulation of the entity and system lists.