Added menu to game transition
The player can start the game by pressing the left, or the right half of the screen. This enables the GameLayer and disables the MenuLayer. The ScrollSystem has to be called before the LevelManagerSystem, otherwise gaps would appear between the blocks. This happens, because new blocks aren't enabled until the next frame, so the ScrollSystem would only move the already existing blocks.