Main menu fade in/out
The main menu fade in/out is achived by seperating the main menu layer in a logic layer and a graphics layer. When the game starts, the logic layer gets disabled and the menu items slowly fade out. The UI is not usable anymore though, because only the logic layer has the button system.